Major Structural Refactor (WIP on Tests)
Still a WIP, I need to fix the tests now. But I've moved from an enum
system to a list/dict system for tracking Bible characteristics. It
seems to make the code cleaner and easier to read. Although I haven't
percluded the possiblity of using a standard class to hold that
information in the long run. It may result in the cleanest code of all.

All that said, this is a WIP until the tests are complete (which is
where the future me should start: boot up the ol' virtual env and run a
good pytest, most of the code should be in the right place - tweak until
all tests pass).

Many of the subroutines were moved into different parent classes. I'm
still trying to determine what makes the most sense, but this next
iteration is a step in the correct direction I believe.

One other major change is the way in references are handled. Now instead
of splitting the result into the start and end verse with
split_reference(), I have a reference_to_verses() method which should do
what split_reference() did, but with several improvements. First it is
now much more flexible: handling multiple books and other extensive
forms of reference input. Second, it now requires a cleaned reference
string to operate - this helped to make it more consistent (or will at
least, once all the bugs are ironed out. *cough* *cough* get on fixing
those TESTS *cough* *hack*) and a corresponding cleanup function was
written to rewrite reference strings, character by character. OOH! That
reminds me, that method has no tests at all yet! Does future me see
where I'm going with this?

TL;DR Dear future me,

FIX THE TESTS. That is all.

Yours truely,

Past Me