Added show() methods for both Verse and Passage
Now there is a way to view the verse content of Passages and Verses.

Also, a bug was fixed where the ESV API would not recognize multi-word
book names with a number in them if the number was rendered
alphabetically. For example, "One Peter" is not recognized but "1 Peter"
is. Now there is a flag in the clean_reference() method allowing
specification of whether or not to convert to human readable (and API
readable) form or leave numbers rendered alphabetically. Alphabetical
rendering is necessary only for the reference_to_verses() method of
Passage.