pytest -> unittest + colour-runner
I've decided to change the testing framework that I use here. In other
parts of my life, I've run into frustrations with the complications and
complexity that pytest introduces, so in my personal projects, I've
decided to avoid it.

I was only using it to run the tests that were defined using unittest
anyway. All I really used it for was to provide colourized test result
output. Don't take on a huge tool just to do something simple I say.
That's where colour-runner is a perfect alternative, all it serves to do
is colourize unittest test output. Perfecto!