Made API Config Methods Static
Now both load_config() and save_config() are static methods within the
API class. This should make it easier to access them without having to
instantiate the entire API object (and load all of the passage objects)
from the tests. Note that I have not refactored the tests yet to reflect
this.