Fixed Fast Mode with Random Single Verse Bug
There was a bug where if you tried to recite a verse in Fast Mode with
Random Single Verse on, the previewed verse at the top of the screen
would always be the first first in the reference, not necessarily the
random verse selected for recitation. This change fixes that.