Implemented Passage & Verse DataClasses
No longer is there a complicated Reference class that does and sees all.
Now there are separated dataclass (read data type) structs that hold the
necessary information.

The Reference dataclass holds information only about the Reference (ie:
1 John 1:1), the Verse contains information only about the content of
the verse (ie: "That which was from the beginning..."), and the Passage
is an object that stores both the Reference and the Verse and (in the
future) other information about the Reference and Verse (like
footnotes/alternate readings?).

Also the agents have been renamed to translations and incorporated with
them. All translations are one-to-one with agents so it made no sense to
distinguish between them.